requirements.txt
asgiref==3.9.1
astpretty==3.0.0
astroid==2.15.8
attrs==25.3.0
cognitive_complexity==1.3.0
Django==5.2.6
django-debug-toolbar==6.0.0
django-phonenumber-field==8.3.0
eradicate==2.3.0
flake8==6.1.0
flake8-blind-except==0.2.1
flake8-builtins==3.0.0
flake8-commas==4.0.0
flake8-comprehensions==3.17.0
flake8-django==1.4
flake8-eradicate==1.5.0
flake8-expression-complexity==0.0.11
flake8-logging-format==2024.24.12
flake8-print==5.0.0
flake8-variables-names==0.0.6
flake8_cognitive_complexity==0.1.0
lazy-object-proxy==1.12.0
mccabe==0.7.0
pep8-naming==0.15.1
phonenumbers==9.0.16
pillow==12.0.0
pycodestyle==2.11.1
pyflakes==3.1.0
setuptools==80.9.0
sqlparse==0.5.3
wrapt==1.17.3
